+++ src/components/new-stuff/index.js
@@ -1,28 +1,23 @@
 import React, { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
-import { withRouter } from 'react-router-dom';
 
 
 import { Overlay, OverlaySection, OverlayTitle, OverlayBackground } from 'Components/overlays';
 import CheckboxButton from 'Components/buttons/checkbox-button';
 import Button from 'Components/buttons/button';
-import PreviewContainer from 'Components/containers/preview-container';
-import Link from 'Components/link';
 import { PopoverContainer } from 'Components/popover';
 
-import { useAPI } from 'State/api';
-import { useGlobals } from 'State/globals';
 import { useTracker } from 'State/segment-analytics';
 import { useCurrentUser } from 'State/current-user';
 import useUserPref from 'State/user-prefs';
 
-import pupdate from '../../curated/pupdate.json';
+import pupdates from '../../curated/pupdates.json';
 import NewStuffArticle from './new-stuff-article';
 import NewStuffPrompt from './new-stuff-prompt';
 import NewStuffPup from './new-stuff-pup';
 import styles from './styles.styl';
 
-const pupdatesArray = pupdate.pupdates;
+const pupdatesArray = pupdates.pupdates;
 const latestId = Math.max(...pupdatesArray.map(({ id }) => id));
 
 function usePreventTabOut() {
@@ -49,7 +44,7 @@ function usePreventTabOut() {
   return { first, last };
 }
 
-const NewStuffOverlay = ({ setShowNewStuff, showNewStuff, newStuff, closePopover }) => {
+export const NewStuffOverlay = ({ setShowNewStuff, showNewStuff, newStuff, closePopover }) => {
   const { first, last } = usePreventTabOut();
 
   return (
@@ -88,33 +83,6 @@ NewStuffOverlay.propTypes = {
     }).isRequired,
   ).isRequired,
 };
-
-export const PupdatePreview = withRouter(({ history }) => {
-  const api = useAPI();
-  const { origin } = useGlobals();
-  const onPublish = async (data) => {
-    try {
-      await api.post(`${origin}/api/home`, data);
-      history.push('/');
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  return (
-    <PreviewContainer
-      get={() => api.get('https://buttercup-room.glitch.me/pupdate.json').then((res) => res.data)}
-      onPublish={onPublish}
-      previewMessage={
-        <>
-          This is a live preview of edits done with the <Link to="https://buttercup-room.glitch.me">Pupdate Editor.</Link>
-        </>
-      }
-    >
-      {(data) => <NewStuffOverlay showNewStuff setShowNewStuff={null} newStuff={data} closePopover={null} />}
-    </PreviewContainer>
-  );
-});
 
 const NewStuff = ({ children }) => {
   const { currentUser } = useCurrentUser();
